Refactor SmallCard component and remove unused code
Remove the 'About Us' case in the SmallCard component and delete the unused handleRouteAboutus function.

Delete unused NotificationImage component

Remove the unused NotificationImage component from the project.

Fix typo in UserCardProps.types.ts

Fix the typo in the import statement for the IUserDetails type in the UserCardProps.types.ts file.

Update profile image in Home component

Update the profile image in the Home component to display the user's profile image if available, or a default image if not.

Remove About Us screen

Delete the About Us screen component as it is no longer needed.